[brightbox] Added check for identifiers being passed to requests to prevent requests to resource root

+++ lib/fog/brightbox/requests/compute/destroy_cloud_ip.rb
@@ -4,6 +4,7 @@ module Fog
       class Real
 
         def destroy_cloud_ip(identifier, options = {})
+          return nil if identifier.nil? || identifier == ""
           request(
             :expects  => [200],
             :method   => 'DELETE',
